+++ jovp/src/main/java/org/lei/opi/jovp/QueryDeserializer.java
@@ -30,12 +30,12 @@ public class QueryDeserializer implements JsonDeserializer<Query> {
         String invGammaFile = jsonObject.get("invGammaFile").getAsString();
         Monitor monitor = context.deserialize(jsonObject.get("monitor"), Monitor.class);
         String webcam = jsonObject.get("webcam").getAsString();
-        double leftEyex = context.deserialize(jsonObject.get("leftEyex"), double.class);
-        double leftEyey = context.deserialize(jsonObject.get("leftEyey"), double.class);
-        double leftEyed = context.deserialize(jsonObject.get("leftEyed"), double.class);
-        double rightEyex = context.deserialize(jsonObject.get("rightEyex"), double.class);
-        double rightEyey = context.deserialize(jsonObject.get("rightEyey"), double.class);
-        double rightEyed = context.deserialize(jsonObject.get("rightEyed"), double.class);
+        int leftEyex = context.deserialize(jsonObject.get("leftEyex"), int.class);
+        int leftEyey = context.deserialize(jsonObject.get("leftEyey"), int.class);
+        int leftEyed = context.deserialize(jsonObject.get("leftEyed"), int.class);
+        int rightEyex = context.deserialize(jsonObject.get("rightEyex"), int.class);
+        int rightEyey = context.deserialize(jsonObject.get("rightEyey"), int.class);
+        int rightEyed = context.deserialize(jsonObject.get("rightEyed"), int.class);
 
         return new Query(distance, fov, viewMode, input, pseudoGray, fullScreen, tracking, maxLum, maxPixel, lumPrecision, invGammaFile, monitor, 
         webcam, leftEyex, leftEyey, leftEyed, rightEyex, rightEyey, rightEyed);
